Arm backend: Add ComputeConstantOpsAOT pass (#9504)
Operators that output tensors based on constant args
are pre-computed and added as buffers.
- The pass currently supports full, arange, linspace, and eye.
- Remove some logic for full now handled by the pass
- Rename FuseConstantOpsPass to FuseConstantArgsPass and do minor
improvements

Fix retracing in FuseViewCopyTransform    
Since the pass can change shapes of ops,
the graph needs to be retraced to show this in node.meta["val"].

+++ backends/arm/tosa_quant_utils.py
@@ -130,7 +130,7 @@ class QuantArgs(NamedTuple):
         ).to(self.dtype)
 
     def dequantize_value(self, qx: torch.Tensor) -> torch.Tensor:
-        return (qx - self.zp) * self.scale
+        return (qx.to(torch.int64) - self.zp) * self.scale
 
     @classmethod
     def from_operator(cls, op, args):
